dockerinit: create a long-running dockerinit
Create a long-running dockerinit which does some setup, starts the
container app, waits for it to finish, and exits with the app's exit
code.

A long-running dockerinit is a prerequisite for using libvirt-lxc, since
libvirt provides no way to return an exit code, and no way to split
stdout and stderr into separate streams.

It also will allow docker to get rid of the ugly Ghost state, and might
even make it possible to do everything in the docker client and get rid
of the docker daemon altogether.