dockerinit: add RPC socket and Resume call
Create a per-container UNIX socket which is used for RPC communication
between the docker daemon and dockerinit.  The socket file is stored in
a bind-mounted directory, seen as /.dockersocket/rpc.sock within the
container.

Also add a Resume() RPC call which will allow docker to pause the
container startup until it's ready to resume.  This will be used in the
future to allow docker to hook up the console before resuming the
container so it doesn't miss any output.  The Resume() call will also be
needed for exit code reporting.